Use save on ctrl s hook
