+++ src/client/src/pages/detail/useSaveOnCtrlS.ts
@@ -1,0 +1,17 @@
+import { useEffect } from "react";
+
+export function useSaveOnKeyDown(callback: () => void) {
+  useEffect(() => {
+    const handleKeyDown = (event: KeyboardEvent) => {
+      if (event.ctrlKey && event.key === "s") {
+        event.preventDefault();
+        callback();
+      }
+    };
+    window.addEventListener("keydown", handleKeyDown);
+
+    return () => {
+      window.removeEventListener("keydown", handleKeyDown);
+    };
+  }, [callback]);
+}
